Add authentication for backend operations (#9)
* Add authentication for backend operations

* limit username

* improve password handling

* check for duplicate users

+++ tests/test_api.py
@@ -21,6 +21,7 @@ from nodeman.x509 import generate_x509_csr
 from tests.utils import CaTestClient
 
 ADMIN_TEST_NODE_COUNT = 100
+BACKEND_CREDENTIALS = ("username", "password")
 
 Settings.model_config = SettingsConfigDict(toml_file="tests/test.toml")
 
@@ -35,12 +36,16 @@ def get_test_client() -> TestClient:
 
 def _test_enroll(data_key, x509_key) -> None:
     client = get_test_client()
+
+    admin_client = get_test_client()
+    admin_client.auth = BACKEND_CREDENTIALS
+
     server = ""
 
     logging.basicConfig(level=logging.DEBUG)
     logging.debug("Testing enrollment")
 
-    response = client.post(urljoin(server, "/api/v1/node"))
+    response = admin_client.post(urljoin(server, "/api/v1/node"))
     assert response.status_code == 201
     create_response = response.json()
     name = create_response["name"]
@@ -49,7 +54,7 @@ def _test_enroll(data_key, x509_key) -> None:
 
     node_url = urljoin(server, f"/api/v1/node/{name}")
 
-    response = client.get(node_url)
+    response = admin_client.get(node_url)
     assert response.status_code == 200
     node_information = response.json()
     assert node_information["name"] == name
@@ -80,7 +85,7 @@ def _test_enroll(data_key, x509_key) -> None:
     response = client.post(node_enroll_url, json=enrollment_request)
     assert response.status_code == 400
 
-    response = client.get(node_url)
+    response = admin_client.get(node_url)
     assert response.status_code == 200
     node_information = response.json()
     print(json.dumps(node_information, indent=4))
@@ -97,10 +102,10 @@ def _test_enroll(data_key, x509_key) -> None:
     assert response.status_code == 200
     _ = load_pem_public_key(response.text.encode())
 
-    response = client.delete(node_url)
+    response = admin_client.delete(node_url)
     assert response.status_code == 204
 
-    response = client.delete(node_url)
+    response = admin_client.delete(node_url)
     assert response.status_code == 404
 
 
@@ -149,7 +154,7 @@ def test_enroll_bad_hmac_signature() -> None:
 
     logging.basicConfig(level=logging.DEBUG)
 
-    response = client.post(urljoin(server, "/api/v1/node"))
+    response = client.post(urljoin(server, "/api/v1/node"), auth=BACKEND_CREDENTIALS)
     assert response.status_code == 201
     create_response = response.json()
     name = create_response["name"]
@@ -174,12 +179,16 @@ def test_enroll_bad_hmac_signature() -> None:
     response = client.post(url, json=enrollment_request)
     assert response.status_code == 401
 
-    response = client.delete(urljoin(server, f"/api/v1/node/{name}"))
+    response = client.delete(urljoin(server, f"/api/v1/node/{name}"), auth=BACKEND_CREDENTIALS)
     assert response.status_code == 204
 
 
 def test_enroll_bad_data_signature() -> None:
     client = get_test_client()
+
+    admin_client = get_test_client()
+    admin_client.auth = BACKEND_CREDENTIALS
+
     server = ""
 
     kty = "OKP"
@@ -187,7 +196,7 @@ def test_enroll_bad_data_signature() -> None:
 
     logging.basicConfig(level=logging.DEBUG)
 
-    response = client.post(urljoin(server, "/api/v1/node"))
+    response = admin_client.post(urljoin(server, "/api/v1/node"))
     assert response.status_code == 201
     create_response = response.json()
     name = create_response["name"]
@@ -214,12 +223,14 @@ def test_enroll_bad_data_signature() -> None:
     response = client.post(url, json=enrollment_request)
     assert response.status_code == 401
 
-    response = client.delete(urljoin(server, f"/api/v1/node/{name}"))
+    response = admin_client.delete(urljoin(server, f"/api/v1/node/{name}"))
     assert response.status_code == 204
 
 
 def test_admin() -> None:
     client = get_test_client()
+    client.auth = BACKEND_CREDENTIALS
+
     server = ""
 
     for _ in range(ADMIN_TEST_NODE_COUNT):
@@ -242,8 +253,30 @@ def test_admin() -> None:
         assert response.status_code == 204
 
 
+def test_backend_authentication() -> None:
+    client = get_test_client()
+    server = ""
+
+    # correct password
+    response = client.get(urljoin(server, "/api/v1/nodes"), auth=BACKEND_CREDENTIALS)
+    assert response.status_code == 200
+
+    # no password
+    response = client.get(urljoin(server, "/api/v1/nodes"))
+    assert response.status_code == 401
+
+    # invalid user
+    response = client.get(urljoin(server, "/api/v1/nodes"), auth=("invalid", ""))
+    assert response.status_code == 401
+
+    # wrong password for existing user
+    response = client.get(urljoin(server, "/api/v1/nodes"), auth=(BACKEND_CREDENTIALS[0], "wrong"))
+    assert response.status_code == 401
+
+
 def test_not_found() -> None:
     client = get_test_client()
+    client.auth = BACKEND_CREDENTIALS
     server = ""
     name = str(uuid.uuid4())
